AW-6026 better class naming and folder structure
